refactor: #35 useApiGetSpaceList 로직 분리
